Module Ring.Front

module Reader : Ring.READABLE

module Writer : Ring.WRITABLE

val write : Cstruct.t -> bytes -> int -> int -> int

write stream buf ofs len writes up to len bytes from buf at ofs to stream. If this returns short it means EOF

val read : Cstruct.t -> bytes -> int -> int -> int

read stream buf ofs len reads up to len bytes to buf at ofs from stream. If this returns short it means EOF

val unsafe_write : Cstruct.t -> bytes -> int -> int -> int

val unsafe_read : Cstruct.t -> bytes -> int -> int -> int