Module Ring

Shared ring handling to communicate with other Xen domains.

type buf = Cstruct.t

module Rpc : sig ... end

module type RW = sig ... end

module Reverse (RW : RW) : RW

module type STREAM = sig ... end

module type READABLE = sig ... end

A stream of readable items

module type WRITABLE = sig ... end

A stream of writable items

module type S = sig ... end

module Pipe (RW : RW) : S

module type Bidirectional_byte_stream = sig ... end

val zero : Cstruct.t -> unit

zero c sets every byte of the c Cstruct to zero.

val unsafe_load_uint32 : Cstruct.t -> int -> int

val unsafe_save_uint32 : Cstruct.t -> int -> int -> unit